rename setting to settings, rename item collection const
